Reorganize Package, Add progress to MainActivity
