--- app/src/main/java/com/jorgesm/themoviedb/Constans.kt
@@ -1,8 +1,0 @@
-package com.jorgesm.themoviedb
-
-
-object Constans {
-
- const val BASE_URL ="https://api.themoviedb.org/3/"
- const val POPULAR_MOVIES_URL = "discover/movie?sort_by=popularity.desc"
-}
