--- app/src/main/java/com/jorgesm/themoviedb/extensions.kt
@@ -1,9 +1,0 @@
-package com.jorgesm.themoviedb
-
-import android.content.Context
-import android.widget.Toast
-
-
-fun Context.toast(message: String){
-    Toast.makeText(this, message, Toast.LENGTH_LONG).show()
-}
