+++ app/src/main/java/com/jorgesm/themoviedb/model/RemoteService.kt
@@ -1,14 +1,14 @@
 package com.jorgesm.themoviedb.model
 
 
-import com.jorgesm.themoviedb.Constans
+import com.jorgesm.themoviedb.utils.Constants
 import retrofit2.http.GET
 import retrofit2.http.Query
 
 
 interface RemoteService {
 
- @GET(Constans.POPULAR_MOVIES_URL)
+ @GET(Constants.POPULAR_MOVIES_URL)
  suspend fun listPopularMovies(
   @Query("api_key") apiKey: String,
   @Query("region") region: String
