+++ app/src/main/java/com/jorgesm/themoviedb/ui/detail/DetailActivity.kt
@@ -1,4 +1,4 @@
-package com.jorgesm.themoviedb.ui
+package com.jorgesm.themoviedb.ui.detail
 
 import android.annotation.SuppressLint
 import android.os.Bundle
