+++ app/src/main/java/com/jorgesm/themoviedb/ui/main/MainActivity.kt
@@ -1,15 +1,17 @@
-package com.jorgesm.themoviedb.ui
+package com.jorgesm.themoviedb.ui.main
 
 
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import android.view.View
 import androidx.lifecycle.lifecycleScope
 import com.jorgesm.themoviedb.databinding.ActivityMainBinding
 import com.jorgesm.themoviedb.utils.Constants
 import kotlinx.coroutines.launch
 import com.jorgesm.themoviedb.model.MoviesRepository
+import com.jorgesm.themoviedb.ui.detail.DetailActivity
 
 class MainActivity : AppCompatActivity() {
     
@@ -30,7 +32,9 @@ class MainActivity : AppCompatActivity() {
         }
         mainBinding.rvCovers.adapter = adapter
         lifecycleScope.launch{
+            mainBinding.progressCircular.visibility = View.VISIBLE
             adapter.submitList( moviesRepository.findPopularMovies().results)
+            mainBinding.progressCircular.visibility = View.GONE
         }
     }
 }
