+++ app/src/main/java/com/jorgesm/themoviedb/utils/Constants.kt
@@ -2,6 +2,9 @@ package com.jorgesm.themoviedb.utils
 
 object Constants {
     const val MOVIE = "DetailActivity:movie"
+    const val BASE_URL ="https://api.themoviedb.org/3/"
+    const val POPULAR_MOVIES_URL = "discover/movie?sort_by=popularity.desc"
+    
     const val IMG_BASE_URL = "https://image.tmdb.org/t/p/"
     const val IMG_780= "w780"
     const val IMG_185= "w185/"
